pyproject.toml
[tool.poetry]
name = "token-validator"
version = "1.0.0"
description = "Lib to validate JWT token"
authors = ["Dilip Sharma <dilip.sharma@anaplan.com>"]
readme = "README.md"
packages = [{ include = "jwt_lib", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12"
pyjwt = ">=2.11.0,<3.0.0"
cryptography = ">=46.0.5"
httpx = ">=0.28.1,<0.29.0"
pytest = { version = ">=9.0.0,<10.0.0", optional = true }
pytest-asyncio = { version = ">=1.3.0,<2.0.0", optional = true }
pytest-cov = { version = ">=4.1.0,<8.0.0", optional = true }
pytest-httpserver = { version = ">=1.0.10,<2.0.0", optional = true }

[tool.poetry.extras]
test = ["pytest", "pytest-asyncio", "pytest-cov", "pytest-httpserver"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
